+++ XingBot/real/elw/RealEvents_k1_4ELW.cs
@@ -25,8 +25,8 @@ namespace XingBot.real
             outData.bidtrad1 = _realDict[szTrCode].GetFieldData("OutBlock", "bidtrad1");
             outData.tradmdvol1 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdvol1"));
             outData.tradmsvol1 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmsvol1"));
-            outData.tradmdrate1 = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdrate1"));
-            outData.tradmsrate1 = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmsrate1"));
+            outData.tradmdrate1 = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdrate1"));
+            outData.tradmsrate1 = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmsrate1"));
             outData.tradmdcha1 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdcha1"));
             outData.tradmscha1 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmscha1"));
 
@@ -36,8 +36,8 @@ namespace XingBot.real
             outData.bidtrad2 = _realDict[szTrCode].GetFieldData("OutBlock", "bidtrad2");
             outData.tradmdvol2 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdvol2"));
             outData.tradmsvol2 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmsvol2"));
-            outData.tradmdrate2 = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdrate2"));
-            outData.tradmsrate2 = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmsrate2"));
+            outData.tradmdrate2 = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdrate2"));
+            outData.tradmsrate2 = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmsrate2"));
             outData.tradmdcha2 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdcha2"));
             outData.tradmscha2 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmscha2"));
 
@@ -47,8 +47,8 @@ namespace XingBot.real
             outData.bidtrad3 = _realDict[szTrCode].GetFieldData("OutBlock", "bidtrad3");
             outData.tradmdvol3 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdvol3"));
             outData.tradmsvol3 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmsvol3"));
-            outData.tradmdrate3 = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdrate3"));
-            outData.tradmsrate3 = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmsrate3"));
+            outData.tradmdrate3 = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdrate3"));
+            outData.tradmsrate3 = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmsrate3"));
             outData.tradmdcha3 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdcha3"));
             outData.tradmscha3 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmscha3"));
 
@@ -58,8 +58,8 @@ namespace XingBot.real
             outData.bidtrad4 = _realDict[szTrCode].GetFieldData("OutBlock", "bidtrad4");
             outData.tradmdvol4 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdvol4"));
             outData.tradmsvol4 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmsvol4"));
-            outData.tradmdrate4 = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdrate4"));
-            outData.tradmsrate4 = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmsrate4"));
+            outData.tradmdrate4 = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdrate4"));
+            outData.tradmsrate4 = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmsrate4"));
             outData.tradmdcha4 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdcha4"));
             outData.tradmscha4 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmscha4"));
 
@@ -69,15 +69,15 @@ namespace XingBot.real
             outData.bidtrad5 = _realDict[szTrCode].GetFieldData("OutBlock", "bidtrad5");
             outData.tradmdvol5 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdvol5"));
             outData.tradmsvol5 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmsvol5"));
-            outData.tradmdrate5 = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdrate5"));
-            outData.tradmsrate5 = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmsrate5"));
+            outData.tradmdrate5 = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdrate5"));
+            outData.tradmsrate5 = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmsrate5"));
             outData.tradmdcha5 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmdcha5"));
             outData.tradmscha5 = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "tradmscha5"));
 
             outData.ftradmdvol = _realDict[szTrCode].GetFieldData("OutBlock", "ftradmdvol");
             outData.ftradmsvol = _realDict[szTrCode].GetFieldData("OutBlock", "ftradmsvol");
-            outData.ftradmdrate = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "ftradmdrate"));
-            outData.ftradmsrate = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "ftradmsrate"));
+            outData.ftradmdrate = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "ftradmdrate"));
+            outData.ftradmsrate = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "ftradmsrate"));
             outData.ftradmdcha = _realDict[szTrCode].GetFieldData("OutBlock", "ftradmdcha");
             outData.ftradmscha = _realDict[szTrCode].GetFieldData("OutBlock", "ftradmscha");
             outData.shcode = _realDict[szTrCode].GetFieldData("OutBlock", "shcode");
