+++ XingBot/real/info/RealEvents_MK2.cs
@@ -23,20 +23,20 @@ namespace XingBot.real
             outData.time = _realDict[szTrCode].GetFieldData("OutBlock", "time");
             outData.kodate = _realDict[szTrCode].GetFieldData("OutBlock", "kodate");
             outData.kotime = _realDict[szTrCode].GetFieldData("OutBlock", "kotime");
-            outData.open = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "open"));
-            outData.high = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "high"));
-            outData.low = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "low"));
-            outData.price = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "price"));
+            outData.open = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "open"));
+            outData.high = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "high"));
+            outData.low = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "low"));
+            outData.price = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "price"));
             outData.sign = _realDict[szTrCode].GetFieldData("OutBlock", "sign");
-            outData.change = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "change"));
-            outData.uprate = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "uprate"));
-            outData.bidho = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "bidho"));
+            outData.change = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "change"));
+            outData.uprate = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "uprate"));
+            outData.bidho = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "bidho"));
             outData.bidrem = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "bidrem"));
-            outData.offerho = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "offerho"));
+            outData.offerho = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "offerho"));
             outData.offerrem = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "offerrem"));
-            outData.volume = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "volume"));
+            outData.volume = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "volume"));
             outData.xsymbol = _realDict[szTrCode].GetFieldData("OutBlock", "xsymbol");
-            outData.cvolume = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "cvolume"));
+            outData.cvolume = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "cvolume"));
         }
     }
 }
