+++ XingBot/real/stock/RealEvents_PH_.cs
@@ -24,7 +24,7 @@ namespace XingBot.real
             outData.sign = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "sign"));
             outData.change = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "change"));
             outData.volume = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "volume"));
-            outData.drate = float.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "drate"));
+            outData.drate = decimal.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "drate"));
             outData.cdhrem = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "cdhrem"));
             outData.cshrem = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "cshrem"));
             outData.bdhrem = long.Parse(_realDict[szTrCode].GetFieldData("OutBlock", "bdhrem"));
